+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0 (October 18th, 2019)
+
+Features:
+
+  - Reworking to match LDAP imports - but with local configuration file
+  - PLEASE NOTE the CONFIGURATION file has changed significantly.
+
 ## 1.2.3 (September 26th, 2018)
 
 Fixes: